+++ packages/react-components/src/lib/file-picker/file-picker.stories.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
-import React, { useCallback, useState } from 'react';
+import React, { useCallback, useState, useMemo, useRef, useEffect } from 'react';
 import { Story } from '@storybook/react/types-6-0';
 import {
   FormControl,
@@ -8,31 +8,326 @@ import {
   Radio,
   RadioGroup,
 } from '@material-ui/core';
+import { Box } from '../box';
 import { SupportedLocales } from '../models';
-import { FileActionData, FilePicker } from './file-picker';
+import { 
+  FileActionData,
+  FilePicker,
+  FileArray,
+  FileData,
+  FileHelper,
+  FilePickerActions
+} from './file-picker';
+import FsMap from './fs-map.json';
 
 export default {
   title: 'File Picker',
   component: FilePicker,
 };
 
-export const ReadOnlyMode: Story = () => <FilePicker readOnlyMode={true} />;
+interface CustomFileData extends FileData {
+  parentId?: string;
+  childrenIds?: string[];
+}
+interface CustomFileMap {
+  [fileId: string]: CustomFileData;
+}
+
+const prepareCustomFileMap = (): Record<string,unknown> => {
+  const baseFileMap = (FsMap.fileMap as unknown) as CustomFileMap;
+  const rootFolderId = FsMap.rootFolderId;
+  return { baseFileMap, rootFolderId };
+};
+
+// Sets up files map and actions
+// eslint-disable-next-line
+const useCustomFileMap = () => {
+  const { baseFileMap, rootFolderId } = useMemo(prepareCustomFileMap, []);
+
+  const [fileMap, setFileMap] = useState<CustomFileMap>(baseFileMap as CustomFileMap);
+  const [currentFolderId, setCurrentFolderId] = useState(rootFolderId);
+
+  const resetFileMap = useCallback(() => {
+    setFileMap(baseFileMap as CustomFileMap);
+    setCurrentFolderId(rootFolderId);
+  }, [baseFileMap, rootFolderId]);
+
+  const currentFolderIdRef = useRef(currentFolderId);
+
+  useEffect(() => {
+    currentFolderIdRef.current = currentFolderId;
+  }, [currentFolderId]);
+
+  const deleteFiles = useCallback((files: CustomFileData[]) => {
+    setFileMap((currentFileMap) => {
+      const newFileMap = { ...currentFileMap };
+
+      files.forEach((file) => {
+        delete newFileMap[file.id];
+
+        if (file.parentId) {
+          const parent = newFileMap[file.parentId];
+          const newChildrenIds = parent.childrenIds?.filter(
+            (id) => id !== file.id
+          );
+          newFileMap[file.parentId] = {
+            ...parent,
+            childrenIds: newChildrenIds,
+            childrenCount: newChildrenIds?.length,
+          };
+        }
+      });
+
+      return newFileMap;
+    });
+  }, []);
+
+  const moveFiles = useCallback(
+    (
+      files: CustomFileData[],
+      source: CustomFileData,
+      destination: CustomFileData
+    ) => {
+      setFileMap((currentFileMap) => {
+        const newFileMap = { ...currentFileMap };
+        const moveFileIds = new Set(files.map((f) => f.id));
+
+        // Delete files from their source folder.
+        const newSourceChildrenIds = source.childrenIds?.filter(
+          (id) => !moveFileIds.has(id)
+        );
+        newFileMap[source.id] = {
+          ...source,
+          childrenIds: newSourceChildrenIds,
+          childrenCount: newSourceChildrenIds?.length,
+        };
+
+        // Add the files to their destination folder.
+        const newDestinationChildrenIds = [
+          ...(destination.childrenIds as string[]),
+          ...files.map((f) => f.id),
+        ];
+        newFileMap[destination.id] = {
+          ...destination,
+          childrenIds: newDestinationChildrenIds,
+          childrenCount: newDestinationChildrenIds.length,
+        };
+
+        // Finally, update the parent folder ID on the files - from source folder
+        // ID to the destination folder ID.
+        files.forEach((file) => {
+          newFileMap[file.id] = {
+            ...file,
+            parentId: destination.id,
+          };
+        });
+
+        return newFileMap;
+      });
+    },
+    []
+  );
+
+  // TODO: in production we should use UUIDs or MD5 hashes for file paths
+  const idCounter = useRef(0);
+  const createFolder = useCallback((folderName: string) => {
+    setFileMap((currentFileMap) => {
+      const newFileMap = { ...currentFileMap };
+
+      // Create the new folder.
+      const newFolderId = `new-folder-${idCounter.current++}`;
+      newFileMap[newFolderId] = {
+        id: newFolderId,
+        name: folderName,
+        isDir: true,
+        modDate: new Date(),
+        parentId: currentFolderIdRef.current,
+        childrenIds: [],
+        childrenCount: 0,
+      };
+
+      // Update parent folder to reference the new folder.
+      const parent = newFileMap[currentFolderIdRef.current];
+      newFileMap[currentFolderIdRef.current] = {
+        ...parent,
+        childrenIds: [...(parent.childrenIds as string[]), newFolderId],
+      };
+
+      return newFileMap;
+    });
+  }, []);
+
+  return {
+    fileMap,
+    currentFolderId,
+    setCurrentFolderId,
+    resetFileMap,
+    deleteFiles,
+    moveFiles,
+    createFolder,
+  };
+};
+
+const useFiles = (
+  fileMap: CustomFileMap,
+  currentFolderId: string
+): FileArray => {
+  return useMemo(() => {
+    const currentFolder = fileMap[currentFolderId];
+    const files = currentFolder.childrenIds
+      ? currentFolder.childrenIds.map(
+          (fileId: string) => fileMap[fileId]
+        )
+      : [];
+    return files;
+  }, [currentFolderId, fileMap]);
+};
+
+const useFolderChain = (
+  fileMap: CustomFileMap,
+  currentFolderId: string
+): FileArray => {
+  return useMemo(() => {
+    const currentFolder = fileMap[currentFolderId];
+    const folderChain = [currentFolder];
+    let parentId = currentFolder.parentId;
+    // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
+    while (parentId) {
+      const parentFile = fileMap[parentId];
+      // eslint-disable-next-line 
+      if (parentFile) {
+        folderChain.unshift(parentFile);
+        parentId = parentFile.parentId;
+      } else {
+        break;
+      }
+    }
+    return folderChain;
+  }, [currentFolderId, fileMap]);
+};
+
+const useFileActionHandler = (
+  setCurrentFolderId: (folderId: string) => void,
+  deleteFiles: (files: CustomFileData[]) => void,
+  moveFiles: (
+    files: FileData[],
+    source: FileData,
+    destination: FileData
+  ) => void,
+  createFolder: (folderName: string) => void
+): (data: FileActionData) => void => {
+  return useCallback(
+    (data: FileActionData) => {
+      if (data.id === FilePickerActions.OpenFiles.id) {
+        const { targetFile, files } = data.payload;
+        const fileToOpen = targetFile ?? files[0];
+        // eslint-disable-next-line
+        if (fileToOpen && FileHelper.isDirectory(fileToOpen)) {
+          setCurrentFolderId(fileToOpen.id);
+          return;
+        }
+      } else if (data.id === FilePickerActions.DeleteFiles.id) {
+        deleteFiles(data.state.selectedFilesForAction);
+      } else if (data.id === FilePickerActions.MoveFiles.id) {
+        moveFiles(
+          data.payload.files,
+          data.payload.source as FileData,
+          data.payload.destination
+        );
+      } else if (data.id === FilePickerActions.CreateFolder.id) {
+        const folderName = prompt('Provide the name for your new folder:');
+        // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
+        if (folderName) createFolder(folderName);
+      }
+    },
+    [createFolder, deleteFiles, moveFiles, setCurrentFolderId]
+  );
+};
+
+export const ReadOnlyMode: Story = () => {
+  const {
+    fileMap,
+    currentFolderId,
+    setCurrentFolderId,
+    // resetFileMap,
+    deleteFiles,
+    moveFiles,
+    createFolder,
+  } = useCustomFileMap();
+  const files = useFiles(fileMap, currentFolderId as string);
+  const folderChain = useFolderChain(fileMap, currentFolderId as string);
+  const handleFileAction = useFileActionHandler(
+    setCurrentFolderId,
+    deleteFiles,
+    moveFiles,
+    createFolder
+  );
+  return (
+    <Box style={{height: '600px'}}>
+      <FilePicker 
+        readOnlyMode={true} 
+        files={files}
+        folderChain={folderChain}
+        onFileAction={handleFileAction}
+      />
+    </Box>
+  );
+};
 
 export const DarkTheme: Story = () => {
+  const {
+    fileMap,
+    currentFolderId,
+    setCurrentFolderId,
+    // resetFileMap,
+    deleteFiles,
+    moveFiles,
+    createFolder,
+  } = useCustomFileMap();
+  const files = useFiles(fileMap, currentFolderId as string);
+  const folderChain = useFolderChain(fileMap, currentFolderId  as string);
+  const handleFileAction = useFileActionHandler(
+    setCurrentFolderId,
+    deleteFiles,
+    moveFiles,
+    createFolder
+  );
   return (
-    <FilePicker
-      theme={{
-        primary: 'blue',
-        background: 'black',
-        surface: 'gray',
-        textOnBackground: 'white',
-        selectionBackground: '#455570',
-      }}
-    />
+    <Box style={{height: '600px'}}>
+      <FilePicker
+        theme={{
+          primary: 'blue',
+          background: 'black',
+          surface: 'gray',
+          textOnBackground: 'white',
+          selectionBackground: '#455570',
+        }}
+        files={files}
+        folderChain={folderChain}
+        onFileAction={handleFileAction}
+      />
+    </Box>
   );
 };
 
 export const Localized: Story = () => {
+  const {
+    fileMap,
+    currentFolderId,
+    setCurrentFolderId,
+    // resetFileMap,
+    deleteFiles,
+    moveFiles,
+    createFolder,
+  } = useCustomFileMap();
+  const files = useFiles(fileMap, currentFolderId as string);
+  const folderChain = useFolderChain(fileMap, currentFolderId  as string);
+  const handleFileAction = useFileActionHandler(
+    setCurrentFolderId,
+    deleteFiles,
+    moveFiles,
+    createFolder
+  );
   const [locale, setLocale] = useState<SupportedLocales>(SupportedLocales.HE);
   const handleLocaleChange = useCallback(
     (event) => setLocale(event.target.value),
@@ -66,14 +361,14 @@ export const Localized: Story = () => {
         </RadioGroup>
       </FormControl>
       <br />
-      <FilePicker locale={locale} />
+      <Box style={{height: '600px'}}>
+        <FilePicker 
+          locale={locale} 
+          files={files}
+          folderChain={folderChain}
+          onFileAction={handleFileAction}
+        />
+      </Box>
     </>
   );
-};
-
-export const CustomFileActionHandler: Story = () => {
-  const handleFileAction = useCallback((data: FileActionData) => {
-    console.log('FileActionData=>', data);
-  }, []);
-  return <FilePicker onFileAction={handleFileAction} />;
 };
